// lib/types.ts
export interface UserJwtPayload {
    userId: string;
}

export interface UserRecord {
    email: string;
    name: string;
    phone?: string;
    address?: string;
    usertype: number;
    shopid?: ShopBrief[];
    password?: string;
    tobeowner: boolean;
    created: string;
    tobeownerAt?: string;
}
// ...
export interface ShopDelete{
    shop_delete: number;
}

export interface NoAccount{
    no_account: number;
}

export interface PasswdErr{
    password_error: number;
}

export interface NoAuthorization{
    no_authorization: number;
}
// ...
export interface ProcessResult{
    status: string;
    message: string;
}
// ...
export interface ShopBrief{
    id: string;
    shopname: string;
    onboard: boolean;
}